separate viewing and editing into different components
